Revert "Zap uint16 t for attribute store (#6446)"
This reverts commit 30daae2b38d8c07c5797f073942113b774f8f299.

+++ src/app/clusters/ias-zone-server/ias-zone-server.cpp
@@ -211,7 +211,7 @@ static void enrollWithClient(EndpointId endpoint)
 }
 
 EmberAfStatus emberAfIasZoneClusterServerPreAttributeChangedCallback(EndpointId endpoint, AttributeId attributeId,
-                                                                     EmberAfAttributeType attributeType, uint16_t size,
+                                                                     EmberAfAttributeType attributeType, uint8_t size,
                                                                      uint8_t * value)
 {
     uint8_t i;
